refactoring all repeated function bodies to the EVMAuthenticator super class
