+++ src/pages/evm/wallet/SendPage.vue
@@ -3,7 +3,7 @@ import { defineComponent } from 'vue';
 import AppPage from 'components/evm/AppPage.vue';
 import UserInfo from 'components/evm/UserInfo.vue';
 import { CURRENT_CONTEXT, getAntelope, useAccountStore, useChainStore, useUserStore } from 'src/antelope';
-import { TransactionResponse, TokenClass, TokenBalance, NativeCurrencyAddress, AntelopeError } from 'src/antelope/types';
+import { TransactionResponse, TokenClass, TokenBalance, NativeCurrencyAddress, AntelopeError, addressString } from 'src/antelope/types';
 import { WEI_PRECISION, formatWei, prettyPrintBalance, prettyPrintFiatBalance } from 'src/antelope/stores/utils';
 import { BigNumber, ethers } from 'ethers';
 import CurrencyInput from 'components/evm/inputs/CurrencyInput.vue';
@@ -242,7 +242,7 @@ export default defineComponent({
 
             const token = this.token as TokenClass;
             const amount = this.amount;
-            const to = this.address;
+            const to = this.address as addressString;
 
             if (this.isFormValid) {
                 ant.stores.balances.transferTokens(token, to, amount).then((trx: TransactionResponse) => {
